Raise exception for processes that terminate with non-zero exit code
Prior to this commit, `core.run` would wait for results indefinitely
when a child process terminated with a non-zero exit code (e.g., segfault).
This was due to the fact that `multiprocessing.Pool` does not raise
an exception ( https://github.com/python/cpython/issues/66587 ). This commit
replaces `multiprocessing.Pool` with `concurrent.futures.ProcessPoolExecutor`.
In contrast to the former, the latter raises an exception in case a
child process terminates with a non-zero exit code.